ServiceWorker scope documentation in explainer.md
The current documentation is a little misleading – it appears to say the default scope is the web root. 
In reality the default scope is relative to the location of the ServiceWorker script.
If a ServiceWorker is located in: `/my-app/sw.js` and the scope is set to: `scope: '/'` an error will throw:

`DOMException: Failed to register a ServiceWorker: The path of the provided scope ('/') is not under the max scope allowed ('/assets/'). Adjust the scope, move the Service Worker script, or use the Service-Worker-Allowed HTTP header to allow the scope.`

I've updated the documentation to show this.