Remove "Doom Installing Worker" algorithm
The algorithm was there for allowing a job queued next to the currently running
register job to replace it right away when the script url is different. However,
there has been a consensus to remove it since this is rather an edge case that
doesn't seem to be paricularly perf sensitive and adds complexity. Fixes #847.