Create a new ReadableStream in respondWith functions
The current description pretends that a ReadableStream can be passed from one
realm to another, which is wrong. This change changes the description so that
it creates a new ReadableStream and passes Uint8Array objects from the old
ReadableStream to the new one.

This is discussed at whatwg/fetch#330. This change fixes #850 as well.