Remove registration's uninstalling flag checking in Unregister
The step that this patch removes is a bit that was used before we
introduced the single job queue that serializes
Register/Update/Unregister jobs. By using the single job queue, the
uninstalling flag set in the Unregister cannot be cleared until the
Unregister job completes. So, we remove this unnecessary step.

Fixes #855.