Change Register algorithm: remove active worker check
Before this patch, Register algorithm returned early (i.e. no-op) only
when the newest worker (that has the same script url as that of the
job) is an active worker. In the f2f, we confirmed that the intention
is to make Register no-op if the script urls are the same regardless
of the state of the newest worker. So, in this patch, we remove the
step that is checking whether the newest worker is an active worker.

Fixes #856.