Change getRegistrations() behavior
Before this patch, getRegistrations() returned all the registrations
in the calling client's origin including those that are in the
uninstalling state. This might be an oversight considering
getRegistration(url) hides the uninstalling registration.
(getRegistration(url) uses the Match Service Worker Registration
algorithm which filters an uninstalling registration.) So, in this
patch, we change the behavior of getRegistrations() to return only the
registrations that are not in uninstalling state.

Fixes #943.