Clarify Clients.matchAll() with active document
Before this patch, Clients.matchAll() was creating/adding WindowClient
objects to the result array even if the corresponding service worker
client is not an active document for its window's browsing context.
This patch adds a condition that checks the browsing context has not
been discarded and the given client's responsible document is the
browsing context's active document to filter out the back-forward
cache pages and unloaded tabs.

Fixes #888.