Correct handling of ExtendableEvent's extend lifetime promises
Before this patch, the algorithm used Promise's wait for all language
to wait for the event's extend lifetime promises to settle. This
wasn't correct as it actually intends to wait for all the promises in
the event's extend lifetime promises to settle before allowing the UA
to terminate the SW. This patch fixes the step by replacing "wait for
all" with "wait until all the promise".

Fixes #936.