Change behavior of skipWaiting()
Before this, skipWaiting() promises when called multiple times on a
waiting worker had different behaviors among each other where the winner
waits unitl the Activate is complete while others don't. This makes the
behavior consistent by making the Activate be called only once. To
achieve that, this removes the steps in skipWaiting() that call the
Activate on a waiting worker (so promise resolves right away after
setting the skip waiting flag) and simplifies the call sites of the
Activate in the Install algorithm.

Related issue: #1015.