Fix promise handling; remove extensions allowed flag
This makes waitUntil() wait only the given promises instead of all the
promises in the given promises' chains. So, it removes the direct
references to the promises internal slots.

This replaces the extensions allowed flag with the combinations of the
pending promises count and the state of the event's dispatch flag.