Editorial: Simplify `ready` (#1136)
(This change doesn't affect any behavior.)

The previous structure for the `ready` algorithm involved a recursive
path through an "else" branch. This tended to obscure the fact that the
corresponding branch was *not* recursive and that the following steps
were indeed reachable.

Re-factor the algorithm to include fewer total branches and to contain
the recursion more tightly.