Fix the type of the key of the registration map
This changes the type of the key of the scope to registration map from a
URL to a string. The algorithms that reference the scope to registration
map (including the Match Service Worker Registration) were correctly
using the keys as strings.

This changes the variable names in the Match Service Worker Registration
to clarify the types of the values of the variables.

(This doesn't affect any behavior.)

Fixes #1118.