Clean up Clients interface methods (#1332)
As a follow-up of the decision to not expose the reserved client, this change
added a filter exclude discarded client. This change also removes the
unnecessary branch that checks whether the environment's creation URL is
trustworthy as now we're dealing with only the clients that are execution ready.
(i.e. the clients must be an environment settings object at least.)

Related issue: https://github.com/w3c/ServiceWorker/issues/1245.