Don't fail update upon error fetching the incumbent worker's import s… (#1377)
* Don't fail update upon error fetching the incumbent worker's import scripts.

To include imported scripts in the byte-for-byte comparison, the Update
algorithm fetches the incumbent worker's scripts in order to
detect if there is a difference. But it would abort the update upon
MIME type error if fetching the script failed. This is undesirable
because the new worker might not use these scripts at all, e.g.,
the scripts may have legitimately been removed from the server.

I considered setting the Updated flag on failure, but it turns out Firefox
treats the failure as "no change" and Chrome's WIP implementation
was also doing that. The consensus at #1374 is to just consider
network errors to be "no change", so do that.