Clarify determination of active service worker and spec plan (#1455)
This change makes it clear that this behavior needs to be (and will be)
specified normatively in HTML Standard, and any future changes for Service
Workers will be incorporated into Service Workers Nightly. It also clarifies,
non-normatively, the behavior of how a service worker client's active service
worker is determined.

Issue: https://github.com/w3c/ServiceWorker/issues/765.
This moves the issue to the next milestone.