Use ServiceWorkerGlobalScope instead of request's settingsObject for COEP. (#11)
When we designed the ServiceWorker static routing API, it was designed to be offloading simple things ServiceWorkers do. With that concept, if we evaluate the cache COEP, it should not be a request's COEP, but to be a ServiceWorker's COEP to make it behave as an offload.

To make it happen, we set up the ServiceWorkerGlobalScope to be used for the COEP check.
Since we only need ServiceWorkerGlobalScope for CSP checks, this PR separated the Run ServiceWorker
algorithm to ServiceWorkerGlobalScope setup and others, and make it called before the static routing's
cache COEP check if ServiceWorkerGlobalScope is not ready.