Add `race-network-and-fetch-handler` source
race-network-and-fetch-handler is a source to allow request to do competition between a network fetch and a fetch handler.  To implement the feature, this CL factored out the process to run a fetch handler in agent.  It is called from both the ServiceWorker static routing API part and the original part.  In the static routing API part, a fetch controller also runs in parallel, and the Handle Fetch algorithm returns a response coming earlier to the queue.