Update editors
Thanks Jake and Marijn for their great contributions to the
ServiceWorker.  Especially, they adviced me on standardizing the
declarative routing (aka. SW static routing API), which brought me lots
of learnings to the ServiceWorker and the standardization process.

Recently, I and Monica became editors, and this change updates the
editor lines.