ServiceWorkerAutoPreload (#1756)
ServiceWorkerAutoPreload is defined as an optional browser optimization so that the browser minimizes the cost of the bootstrap by dispatching the navigational network request before starting the service worker. The dispatched network request is consumed by fetch(e.request) in the fetch handler after the bootstrap, or consumed by the client as a fallback request if the fetch handler doesn't call respondWith().

The developer can disable this optimization by opting out via setting the router rule provided by the static routing API. Specifically, the router source "network" works as an opt-out signal.